fix iter
